+++ tidy3d/components/tcad/boundary/generic.py
@@ -1,0 +1,3 @@
+HeatChargeBoundaryConditionType = Union[
+    TemperatureBC, HeatFluxBC, ConvectionBC, VoltageBC, CurrentBC, InsulatingBC
+]
